factor out globus group error component
clean up
